hoists Extension context to DesignPage
