+++ Composer/packages/client/src/pages/design/PropertyEditor.tsx
@@ -3,18 +3,18 @@
 
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
-import React, { useEffect, useMemo, useRef, useState } from 'react';
+import React, { useEffect, useMemo, useRef, useState, useContext } from 'react';
 import AdaptiveForm, { resolveRef, getUIOptions } from '@bfc/adaptive-form';
-import Extension, { FormErrors, JSONSchema7, PluginConfig } from '@bfc/extension';
+import { FormErrors, JSONSchema7 } from '@bfc/extension';
 import formatMessage from 'format-message';
 import isEqual from 'lodash/isEqual';
 import debounce from 'lodash/debounce';
 import mapValues from 'lodash/mapValues';
 import { Resizable, ResizeCallback } from 're-resizable';
 import { MicrosoftAdaptiveDialog } from '@bfc/shared';
+import ExtensionContext from '@bfc/extension/lib/extensionContext';
 
 import { useShell } from '../../shell';
-import plugins, { mergePluginConfigs } from '../../plugins';
 
 import { formEditor } from './styles';
 
@@ -57,18 +57,13 @@ const PropertyEditor: React.FC = () => {
     };
   }, [formData]);
 
+  const { plugins: pluginConfig } = useContext(ExtensionContext);
+
   const $schema = useMemo(() => {
     if (schemas?.sdk?.content && localData) {
       return resolveBaseSchema(schemas.sdk.content, localData.$kind);
     }
   }, [schemas?.sdk?.content, localData.$kind]);
-
-  const pluginConfig: PluginConfig = useMemo(() => {
-    const sdkUISchema = schemas?.ui?.content ?? {};
-    const userUISchema = schemas?.uiOverrides?.content ?? {};
-
-    return mergePluginConfigs({ uiSchema: sdkUISchema }, plugins, { uiSchema: userUISchema });
-  }, [schemas?.ui?.content, schemas?.uiOverrides?.content]);
 
   const $uiOptions = useMemo(() => {
     return getUIOptions($schema, mapValues(pluginConfig.uiSchema, 'form'));
@@ -133,15 +128,13 @@ const PropertyEditor: React.FC = () => {
       onResizeStop={handleResize}
     >
       <div aria-label={formatMessage('form editor')} css={formEditor} data-testid="PropertyEditor" role="region">
-        <Extension plugins={pluginConfig} shell={shellApi} shellData={shellData}>
-          <AdaptiveForm
-            errors={errors}
-            formData={localData}
-            schema={$schema}
-            uiOptions={$uiOptions}
-            onChange={handleDataChange}
-          />
-        </Extension>
+        <AdaptiveForm
+          errors={errors}
+          formData={localData}
+          schema={$schema}
+          uiOptions={$uiOptions}
+          onChange={handleDataChange}
+        />
       </div>
     </Resizable>
   );
